final submission of the project

+++ C0747798_Mid_Term_MAD3004/Product .swift
@@ -11,19 +11,19 @@ class Product: Manufacturer
 {
     var Product_Id: String
     var Product_Name: String
-    var Price: Float
+    var Price: Double
     var Quantity: Int
 
 override init()
 {
     Product_Id = String()
     Product_Name = String()
-    Price = Float()
+    Price = Double()
     Quantity = Int()
     super.init()
 }
     
-init(ProId:String, ProName:String, Price:Float, Quantity: Int)
+init(ProId:String, ProName:String, Price:Double, Quantity: Int)
     {
         self.Product_Id = ProId
         self.Product_Name = ProName
